Moved views from core to custom module
