+++ corehq/apps/hqwebapp/models.py
@@ -341,7 +341,6 @@ class CommTrackSetupTab(UITab):
             NewProgramView,
             EditProgramView,
             SMSSettingsView,
-            ILSConfigView,
         )
         from corehq.apps.locations.views import (
             LocationsListView,
@@ -353,9 +352,8 @@ class CommTrackSetupTab(UITab):
             LocationSettingsView,
         )
 
-        items = []
 
-        items.append([_('CommTrack Setup'), [
+        return [[_('CommTrack Setup'), [
             # products
             {
                 'title': ProductListView.page_title,
@@ -434,13 +432,7 @@ class CommTrackSetupTab(UITab):
                 'title': FacilitySyncView.page_title,
                 'url': reverse(FacilitySyncView.urlname, args=[self.domain]),
             },
-        ]])
-        if self.couch_user and (self.couch_user.is_superuser or IS_DEVELOPER.enabled(self.couch_user.username)):
-            items[0][1].append({
-                'title': ILSConfigView.page_title,
-                'url': reverse(ILSConfigView.urlname, args=[self.domain]),
-            })
-        return items
+        ]]]
 
 
 class ProjectDataTab(UITab):
