+++ custom/ilsgateway/urls.py
@@ -1,7 +1,10 @@
 from django.conf.urls import patterns, url
 from custom.ilsgateway.views import GlobalStats
-
+from custom.ilsgateway.views import ILSConfigView
 
 urlpatterns = patterns('custom.ilsgateway.views',
+    url(r'^ils_config/$', ILSConfigView.as_view(), name=ILSConfigView.urlname),
+    url(r'^sync_ilsgateway/$', 'sync_ilsgateway', name='sync_ilsgateway'),
     url(r'^global_stats/$', GlobalStats.as_view(), name=GlobalStats.urlname)
+
 )
